+++ public/assets/icons/index.ts
@@ -12,7 +12,9 @@ export { default as IconDropdown } from "./icon-dropdown.svg";
 export { default as IconGear } from "./icon-gear.svg";
 export { default as IconImgSelect } from "./icon-img-select.svg";
 export { default as IconPlus } from "./icon-plus.svg";
+export { default as IconPlusCurrent } from "./icon-plus-current.svg";
 export { default as IconProfile } from "./icon-profile.svg";
+export { default as IconProfileCurrent } from "./icon-profile-current.svg";
 export { default as IconProfileDesktop } from "./icon-profile-desktop.svg";
 export { default as IconToDo } from "./icon-todo.svg";
 export { default as IconToggle } from "./icon-toggle.svg";
